cyber-tulas/manager-password/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.1
fastapi==0.110.0
h11==0.16.0
idna==3.7
motor==3.4.0
pydantic==2.7.1
pymongo==4.7.1
pydantic-core==2.18.2
sniffio==1.3.1
typing-extensions==4.11.0
uvicorn==0.29.0
python-dotenv
pytest
pytest-asyncio
httpx
mongomock
passlib
python-multipart
pyjwt
pydantic[email]
stripe
sqlalchemy
prometheus_fastapi_instrumentator
python-logging-loki==0.3.1